visibility-on-load.html

<!DOCTYPE HTML>
<html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
  assert_equals(document.visibilityState, 'visible');
}, 'document.visibilityState during parsing');

async_test(t => {
  document.addEventListener('DOMContentLoaded', t.step_func_done(() => {
    assert_equals(document.visibilityState, 'visible');
  }));
}, 'document.visibilityState at DOMContentLoaded event');

async_test(t => {
  window.addEventListener('load', t.step_func_done(() => {
    assert_equals(document.visibilityState, 'visible');
  }));
}, 'document.visibilityState at load event');
</script>
